move UnwrapOrRevert to mod
